Refactor from review.
